Added Bottomsheet Under Widgets Folder and Removed AI Generated Comments
